+++ engine/include/physic/plane.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif // __cplusplus
 
-int8_t plane_intersect_plane3(plane_t p1, plane_t p2, plane_t p3, vector3_t *intersection_point);
+uint32_t plane_intersect_plane3(plane_t p1, plane_t p2, plane_t p3, vector3_t *intersection_point);
 
 #ifdef __cplusplus
 }
