+++ engine/include/physic/ray.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif // __cplusplus
 
-int8_t ray_intersect_aabb(ray_t ray, aabb_t aabb, float *t_enter, float *t_exit);
+uint32_t ray_intersect_aabb(ray_t ray, aabb_t aabb, float *t_enter, float *t_exit);
 
 #ifdef __cplusplus
 }
