+++ engine/include/swapchain.h
@@ -4,7 +4,7 @@
 #define SWAPCHAIN_MAX_IMAGE_COUNT (0x10)
 
 typedef struct swapchain_t {
-  int8_t is_dirty;
+  uint32_t is_dirty;
   int32_t image_count;
   VkSwapchainKHR handle;
 } swapchain_t;
