+++ engine/include/window.h
@@ -140,7 +140,7 @@ typedef struct window_t {
   float time;
   float delta_time;
   float elapsed_time_since_fps_count_update;
-  int8_t is_window_running;
+  uint32_t is_window_running;
   int32_t window_width;
   int32_t window_height;
   int32_t mouse_position_x;
@@ -188,13 +188,13 @@ void window_create(int32_t width, int32_t height, char const *title);
 void window_run(void);
 void window_destroy(void);
 
-uint8_t window_is_keyboard_key_pressed(keyboard_key_t key);
-uint8_t window_is_keyboard_key_held(keyboard_key_t key);
-uint8_t window_is_keyboard_key_released(keyboard_key_t key);
+uint32_t window_is_keyboard_key_pressed(keyboard_key_t key);
+uint32_t window_is_keyboard_key_held(keyboard_key_t key);
+uint32_t window_is_keyboard_key_released(keyboard_key_t key);
 
-uint8_t window_is_mouse_key_pressed(mouse_key_t key);
-uint8_t window_is_mouse_key_held(mouse_key_t key);
-uint8_t window_is_mouse_key_released(mouse_key_t key);
+uint32_t window_is_mouse_key_pressed(mouse_key_t key);
+uint32_t window_is_mouse_key_held(mouse_key_t key);
+uint32_t window_is_mouse_key_released(mouse_key_t key);
 
 #ifdef __cplusplus
 }
